test/regression/reftest/vendor/blink/fast/dom/clone-node-default-argument.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/wpt-test-pre.js"></script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

shouldBe('Node.prototype.cloneNode.length', '0');

var div = document.createElement('div');
div.innerHTML = '<a></a>';
shouldBeFalse('div.cloneNode().hasChildNodes()');
shouldBe('div.cloneNode().childNodes.length', '0');

</script>
<script src="../../resources/js-test-post.js"></script>
<script src="../../resources/wpt-test-post.js"></script>
</body>
</html>
